requirements.txt
# Django Packages
Django==5.1
django-autoslug==1.9.9
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-imagekit==5.0.0
django-phonenumber-field==8.0.0
django-tables2==2.7.0

# Bootstrap Packages
crispy-bootstrap5==2024.2

# Utility Packages
openpyxl==3.1.5
phonenumbers==8.13.43
requests==2.32.3
tablib==3.6.1

pillow~=11.0.0
djangorestframework~=3.15.2